+++ src/utils/customCommands.ts
@@ -1,54 +1,90 @@
-import vscode from "vscode";
+import { window, Terminal, TerminalShellExecution } from "vscode";
 
-const customCommands: {
-  [key: string]: (terminal: any, args: string[]) => void;
-} = {
-  "*stop": handleStop,
-  "*close": handleClose,
-  "*open_file": openFile,
-  "*alert": alertMessage,
-  "*echo": echo,
+export type CommandResult =
+  | { type: "execution"; execution: TerminalShellExecution | undefined }
+  | { type: "cancel" }
+  | { type: "continue" }
+  | { type: "error"; error: Error };
+
+type ExecuteCommand = (
+  command: string,
+  terminal: Terminal
+) => TerminalShellExecution | undefined;
+
+type CustomCommand = (
+  terminal: Terminal,
+  args: string[],
+  executeCommand: ExecuteCommand
+) => Promise<CommandResult>;
+
+type CustomCommands = {
+  [key: string]: CustomCommand;
 };
 
-async function handleStop(terminal: vscode.Terminal) {
+const customCommands: CustomCommands = {
+  "*stop": handleStop,
+  "*close": handleClose,
+  "*alert": alertMessage,
+  "*echo": echo,
+  "*delay": handleDelay,
+};
+
+async function handleStop(
+  terminal: Terminal,
+  args: string[],
+  executeCommand: ExecuteCommand
+): Promise<CommandResult> {
   const pid = await terminal.processId;
-  terminal.sendText(`kill -SIGINT ${pid}`);
+  const execution = executeCommand(`kill -SIGINT ${pid}`, terminal);
+  return { type: "execution", execution };
 }
 
-function handleClose(terminal: vscode.Terminal) {
-  terminal.dispose();
+async function handleClose(terminal: Terminal): Promise<CommandResult> {
+  return await new Promise<CommandResult>((resolve) => {
+    const closeTerminalListener = window.onDidCloseTerminal(
+      (closedTerminal) => {
+        if (closedTerminal === terminal) {
+          closeTerminalListener.dispose();
+          resolve({ type: "cancel" });
+        }
+      }
+    );
+
+    terminal.dispose();
+  });
 }
 
-function openFile(terminal: vscode.Terminal, args: string[]) {
-  if (args.length === 0) {
-    vscode.window.showErrorMessage("No file name provided");
-    return;
-  }
-
-  const fileName = args[0];
-  if (
-    !vscode.workspace.textDocuments.some((doc) => doc.fileName === fileName)
-  ) {
-    vscode.window.showErrorMessage(`File ${fileName} does not exist`);
-    return;
-  }
-
-  vscode.workspace
-    .openTextDocument(fileName)
-    .then((doc: vscode.TextDocument) => {
-      vscode.window.showTextDocument(doc);
-    });
-}
-
-function echo(terminal: vscode.Terminal, args: string[]) {
+async function echo(
+  terminal: Terminal,
+  args: string[],
+  executeCommand: ExecuteCommand
+): Promise<CommandResult> {
   const message = args.join(" ");
-  terminal.sendText(`echo "${message}"\n`);
+  const execution = executeCommand(`echo "${message}"\n`, terminal);
+  return { type: "execution", execution };
 }
 
-function alertMessage(terminal: vscode.Terminal, args: string[]) {
-  // vscode only allows 3 notifications open at a time
+async function alertMessage(
+  terminal: Terminal,
+  args: string[]
+): Promise<CommandResult> {
   const message = args.join(" ");
-  vscode.window.showInformationMessage(message);
+  window.showInformationMessage(message);
+  await delay(1000);
+  return { type: "continue" };
+}
+
+async function handleDelay(
+  terminal: Terminal,
+  args: string[]
+): Promise<CommandResult> {
+  const seconds = Number(args[0]);
+  await delay(seconds);
+  return { type: "continue" };
+}
+
+function delay(ms: number) {
+  return new Promise((resolve) => setTimeout(resolve, ms));
 }
 
 export default customCommands;